requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
anndata==0.11.4
array-api-compat==1.12.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blosc2==3.5.1
certifi==2025.6.15
charset-normalizer==3.4.2
cmake==4.0.3
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
einops==0.6.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
igraph==0.11.9
imageio==2.37.0
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lazy_loader==0.4
legacy-api-wrap==1.4.1
leidenalg==0.10.2
lightning-utilities==0.14.3
lit==18.1.8
llvmlite==0.44.0
louvain==0.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
natsort==8.4.0
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numexpr==2.11.0
numpy==1.26.0
#nvidia-cublas-cu11==11.10.3.66
#nvidia-cuda-cupti-cu11==11.7.101
#nvidia-cuda-nvrtc-cu11==11.7.99
#nvidia-cuda-runtime-cu11==11.7.99
#nvidia-cudnn-cu11==8.5.0.96
#nvidia-cufft-cu11==10.9.0.58
#nvidia-curand-cu11==10.2.10.91
#nvidia-cusolver-cu11==11.4.0.1
#nvidia-cusparse-cu11==11.7.4.91
#nvidia-nccl-cu11==2.14.3
#nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.1.78
packaging==25.0
pandas==2.3.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
Pillow==10.0.0
pip==24.2
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-igraph==0.11.9
pytorch-lightning==2.0.8
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scanpy==1.11.3
scikit-fuzzy==0.5.0
scikit-image==0.21.0
scikit-learn==1.7.0
scipy==1.14.0
seaborn==0.13.2
session-info2==0.1.2
setuptools==75.1.0
sinfo==0.3.4
six==1.17.0
stack-data==0.6.3
statsmodels==0.14.5
stdlib-list==0.11.1
sympy==1.14.0
tables==3.10.1
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.9
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.1
torchmetrics==1.7.4
torchvision==0.19.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.2
# triton==2.0.0
typing==3.7.4.3
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.44.0
yarl==1.20.1
harmonypy==0.0.10
brokenaxes==0.6.2
openslide-python